requirements.txt
python-dotenv
anthropic
streamlit